Model cleanup. Separates core data transport stuff to extensions so we can use the view models without a core data dependency.
